backend/requirements.txt
fastapi==0.111.0
python-dotenv==1.0.1
groq>=0.8.0
pydantic==2.8.2
